AsyncSocket: BeginXXX to XXXAsync

+++ Zeze/Net/AsyncSocket.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Net.Sockets;
+using System.Diagnostics;
 
 namespace Zeze.Net
 {
@@ -12,15 +13,16 @@ namespace Zeze.Net
     /// </summary>
     public class AsyncSocket : IDisposable
     {
-        private Socket _socket;
         private Zeze.Serialize.ByteBuffer _inputBuffer;
         private List<System.ArraySegment<byte>> _outputBufferList = null;
+        private int _outputBufferListCountSum = 0;
         private List<System.ArraySegment<byte>> _outputBufferListSending = null; // 正在发送的 buffers.
+        private int _outputBufferListSendingCountSum = 0;
 
         public Manager Manager { get; private set; }
         public Exception LastException { get; private set; }
         public long SerialNo { get; private set; }
-        public Socket Socket {  get { return _socket; } } // 这个给出去真的好吗？
+        public Socket Socket { get; private set; } // 这个给出去真的好吗？
 
         /// <summary>
         /// 保存需要存储在Socket中的状态，比如加解密的功能。
@@ -31,6 +33,10 @@ namespace Zeze.Net
 
         private static Zeze.Util.AtomicLong SerialNoGen = new Zeze.Util.AtomicLong();
 
+        private SocketAsyncEventArgs eventArgsAccept;
+        private SocketAsyncEventArgs eventArgsReceive;
+        private SocketAsyncEventArgs eventArgsSend;
+
         /// <summary>
         /// for server socket
         /// </summary>
@@ -38,43 +44,46 @@ namespace Zeze.Net
         {
             this.Manager = manager;
 
-            _socket = new Socket(SocketType.Stream, ProtocolType.Tcp);
-            _socket.Blocking = false;
-            _socket.SetSocketOption(SocketOptionLevel.Socket, SocketOptionName.ReuseAddress, true);
- 
+            Socket = new Socket(SocketType.Stream, ProtocolType.Tcp);
+            Socket.Blocking = false;
+            Socket.SetSocketOption(SocketOptionLevel.Socket, SocketOptionName.ReuseAddress, true);
+
             // xxx 只能设置到 ServerSocket 中，以后 Accept 的连接通过继承机制得到这个配置。
             // 不知道 c# 会不会也这样，先这样写。
             if (null != manager.SocketOptions.ReceiveBuffer)
-                _socket.ReceiveBufferSize = manager.SocketOptions.ReceiveBuffer.Value;
+                Socket.ReceiveBufferSize = manager.SocketOptions.ReceiveBuffer.Value;
 
-            _socket.Bind(localEP);
-            _socket.Listen(manager.SocketOptions.Backlog);
-            _socket.BeginAccept(manager.SocketOptions.InputBufferInitCapacity, OnAsyncAccept, this);
+            Socket.Bind(localEP);
+            Socket.Listen(manager.SocketOptions.Backlog);
 
             this.SerialNo = SerialNoGen.IncrementAndGet();
+
+            eventArgsAccept = new SocketAsyncEventArgs();
+            eventArgsAccept.Completed += OnAsyncIOCompleted;
+
+            BeginAcceptAsync();
         }
 
         /// <summary>
         /// use inner. create when accept;
         /// </summary>
         /// <param name="accepted"></param>
-        AsyncSocket(Manager manager, Socket accepted, byte[] bytes, int bytesTransferred)
+        AsyncSocket(Manager manager, Socket accepted)
         {
             this.Manager = manager;
 
-            _socket = accepted;
-            _socket.Blocking = false;
+            Socket = accepted;
+            Socket.Blocking = false;
 
             // 据说连接接受以后设置无效，应该从 ServerSocket 继承
             if (null != manager.SocketOptions.ReceiveBuffer)
-                _socket.ReceiveBufferSize = manager.SocketOptions.ReceiveBuffer.Value;
+                Socket.ReceiveBufferSize = manager.SocketOptions.ReceiveBuffer.Value;
             if (null != manager.SocketOptions.SendBuffer)
-                _socket.SendBufferSize = manager.SocketOptions.SendBuffer.Value;
+                Socket.SendBufferSize = manager.SocketOptions.SendBuffer.Value;
             if (null != manager.SocketOptions.NoDelay)
-                _socket.NoDelay = manager.SocketOptions.NoDelay.Value;
+                Socket.NoDelay = manager.SocketOptions.NoDelay.Value;
 
-            _inputBuffer = Zeze.Serialize.ByteBuffer.Wrap(bytes, 0, bytesTransferred);
-            // BeginReceive called in OnAsyncAccept
+            BeginReceiveAsync();
 
             this.SerialNo = SerialNoGen.IncrementAndGet();
         }
@@ -88,18 +97,18 @@ namespace Zeze.Net
         {
             this.Manager = manager;
 
-            _socket = new Socket(SocketType.Stream, ProtocolType.Tcp);
-            _socket.Blocking = false;
-            
+            Socket = new Socket(SocketType.Stream, ProtocolType.Tcp);
+            Socket.Blocking = false;
+
             if (null != manager.SocketOptions.ReceiveBuffer)
-                _socket.ReceiveBufferSize = manager.SocketOptions.ReceiveBuffer.Value;
+                Socket.ReceiveBufferSize = manager.SocketOptions.ReceiveBuffer.Value;
             if (null != manager.SocketOptions.SendBuffer)
-                _socket.SendBufferSize = manager.SocketOptions.SendBuffer.Value;
+                Socket.SendBufferSize = manager.SocketOptions.SendBuffer.Value;
             if (null != manager.SocketOptions.NoDelay)
-                _socket.NoDelay = manager.SocketOptions.NoDelay.Value;
+                Socket.NoDelay = manager.SocketOptions.NoDelay.Value;
 
             System.Net.Dns.BeginGetHostAddresses(host, OnAsyncGetHostAddresses, port);
- 
+
             this.SerialNo = SerialNoGen.IncrementAndGet();
         }
 
@@ -121,13 +130,18 @@ namespace Zeze.Net
         /// <param name="length"></param>
         public void Send(byte[] bytes, int offset, int length)
         {
+            Zeze.Serialize.Helper.VerifyParameter(bytes, offset, length);
+            if (length <= 0)
+                return; // 不允许发送空缓存
+
             lock (this)
             {
                 if (null == _outputBufferList)
                     _outputBufferList = new List<ArraySegment<byte>>();
                 _outputBufferList.Add(new ArraySegment<byte>(bytes, offset, length));
+                _outputBufferListCountSum += _outputBufferList[^1].Count;
             }
-            BeginSend(0); // must be zero.
+            BeginSendAsync(0); // must be zero.
         }
 
         public void Send(string str)
@@ -135,29 +149,44 @@ namespace Zeze.Net
             Send(Encoding.UTF8.GetBytes(str));
         }
 
-        private void OnAsyncAccept(IAsyncResult ar)
+        private void OnAsyncIOCompleted(object sender, SocketAsyncEventArgs e)
         {
+            switch (e.LastOperation)
+            {
+                case SocketAsyncOperation.Accept:
+                    ProcessAccept(e);
+                    break;
+                case SocketAsyncOperation.Send:
+                    ProcessSend(e);
+                    break;
+                case SocketAsyncOperation.Receive:
+                    ProcessReceive(e);
+                    break;
+                default:
+                    throw new ArgumentException();
+            }
+        }
+
+        private void BeginAcceptAsync()
+        {
+            eventArgsAccept.AcceptSocket = null;
+            if (false == Socket.AcceptAsync(eventArgsAccept))
+                ProcessAccept(eventArgsAccept);
+        }
+
+        private void ProcessAccept(SocketAsyncEventArgs e)
+        {
+            AsyncSocket accepted = null;
             try
             {
-                byte[] buffer;
-                int bytesTransferred;
-                System.Net.Sockets.Socket so = _socket.EndAccept(out buffer, out bytesTransferred, ar);
-                AsyncSocket aso = new AsyncSocket(this.Manager, so, buffer, bytesTransferred);
-                try
-                {
-                    this.Manager.OnSocketAccept(aso);
-                    this.Manager.OnSocketProcessInputBuffer(aso, aso._inputBuffer);
-                    aso.BeginReceive();
-                }
-                catch (Exception ce)
-                {
-                    aso.Close(ce);
-                }
+                accepted = new AsyncSocket(this.Manager, e.AcceptSocket);
+                this.Manager.OnSocketAccept(accepted);
             }
-            catch (Exception e)
+            catch (Exception ce)
             {
-                Close(e);
+                accepted?.Close(ce);
             }
+            BeginAcceptAsync();
         }
 
         private void OnAsyncGetHostAddresses(IAsyncResult ar)
@@ -166,7 +195,7 @@ namespace Zeze.Net
             {
                 int port = (System.Int32)ar.AsyncState;
                 System.Net.IPAddress[] addrs = System.Net.Dns.EndGetHostAddresses(ar);
-                _socket.BeginConnect(addrs, port, OnAsyncConnect, this);
+                Socket.BeginConnect(addrs, port, OnAsyncConnect, this);
             }
             catch (Exception e)
             {
@@ -179,9 +208,9 @@ namespace Zeze.Net
         {
             try
             {
-                this._socket.EndConnect(ar);
-                BeginReceive();
+                this.Socket.EndConnect(ar);
                 this.Manager.OnSocketConnected(this);
+                BeginReceiveAsync();
             }
             catch (Exception e)
             {
@@ -190,8 +219,14 @@ namespace Zeze.Net
             }
         }
 
-        private void BeginReceive()
+        private void BeginReceiveAsync()
         {
+            if (null == eventArgsReceive)
+            {
+                eventArgsReceive = new SocketAsyncEventArgs();
+                eventArgsReceive.Completed += OnAsyncIOCompleted;
+            }
+
             if (null == _inputBuffer)
             {
                 _inputBuffer = Serialize.ByteBuffer.Allocate(Manager.SocketOptions.InputBufferInitCapacity);
@@ -200,7 +235,7 @@ namespace Zeze.Net
             {
                 if (_inputBuffer.Size == 0 && _inputBuffer.Capacity > Manager.SocketOptions.InputBufferResetThreshold)
                     _inputBuffer = Serialize.ByteBuffer.Allocate(Manager.SocketOptions.InputBufferInitCapacity);
-                _inputBuffer.Campact();
+                _inputBuffer.Campact(); // 上次接收还有剩余数据.
             }
 
             if (_inputBuffer.Capacity - _inputBuffer.WriteIndex < Manager.SocketOptions.InputBufferInitCapacity)
@@ -209,40 +244,53 @@ namespace Zeze.Net
             byte[] buffer = _inputBuffer.Bytes;
             int offset = _inputBuffer.WriteIndex;
             int size = _inputBuffer.Capacity - _inputBuffer.WriteIndex;
-            this._socket.BeginReceive(buffer, offset, size, System.Net.Sockets.SocketFlags.None, OnAsyncBeginReceive, this);
+            eventArgsReceive.SetBuffer(buffer, offset, size);
+
+            if (false == this.Socket.ReceiveAsync(eventArgsReceive))
+                ProcessReceive(eventArgsReceive);
         }
 
-        private void OnAsyncBeginReceive(IAsyncResult ar)
+        private void ProcessReceive(SocketAsyncEventArgs e)
         {
-            try
+            if (e.BytesTransferred > 0 && e.SocketError == SocketError.Success)
             {
-                int received = this._socket.EndReceive(ar);
-                if (received > 0)
+                _inputBuffer.WriteIndex += e.BytesTransferred;
+                if (_inputBuffer.WriteIndex > _inputBuffer.Capacity) // 这个应该不会发生。
                 {
-                    _inputBuffer.WriteIndex += received;
-                    if (_inputBuffer.WriteIndex > _inputBuffer.Capacity) // 这个应该不会发生。
-                    {
-                        Close(new Exception("input buffer overflow."));
-                        return;
-                    }
-                    this.Manager.OnSocketProcessInputBuffer(this, _inputBuffer);
-                    BeginReceive();
+                    Close(new Exception("input buffer overflow."));
+                    return;
                 }
-                else
-                {
-                    Close(null); // 正常关闭，不设置异常
-                }
+                this.Manager.OnSocketProcessInputBuffer(this, _inputBuffer);
+                BeginReceiveAsync();
             }
-            catch (Exception e)
+            else
             {
-                Close(e);
+                Close(null); // 正常关闭，不设置异常
             }
         }
 
-        private void BeginSend(int hasSend)
+        private void ProcessSend(SocketAsyncEventArgs e)
+        {
+            if (e.BytesTransferred > 0 && e.SocketError == SocketError.Success)
+            {
+                BeginSendAsync(e.BytesTransferred); // 必须大于0
+            }
+            else
+            {
+                Close(new SocketException((int)e.SocketError)) ;
+            }
+        }
+
+        private void BeginSendAsync(int hasSend)
         {
             lock(this)
             {
+                if (null == eventArgsSend)
+                {
+                    eventArgsSend = new SocketAsyncEventArgs();
+                    eventArgsSend.Completed += OnAsyncIOCompleted;
+                }
+
                 if (hasSend > 0)
                 {
                     // 听说 BeginSend 成功回调的时候，所有数据都会被发送，这样的话就可以直接清除_outputBufferSending，而不用这么麻烦。
@@ -250,41 +298,57 @@ namespace Zeze.Net
                     // _outputBufferSending != null
                     // _outputBufferSending.Count > 0
                     // sum(_outputBufferSending[i].Count) <= hasSend
-                    for (int i = 0; i < _outputBufferListSending.Count; ++i)
+                    if (hasSend == _outputBufferListSendingCountSum) // 全部发送完
                     {
-                        int bytesCount = _outputBufferListSending[i].Count;
-                        if (hasSend >= bytesCount)
+                        _outputBufferListSending.Clear();
+                    }
+                    else if (hasSend > _outputBufferListSendingCountSum)
+                    {
+                        throw new Exception("hasSend too big.");
+                    }
+                    else
+                    {
+                        for (int i = 0; i < _outputBufferListSending.Count; ++i)
                         {
-                            hasSend -= bytesCount;
-                            if (hasSend > 0)
-                                continue;
+                            int bytesCount = _outputBufferListSending[i].Count;
+                            if (hasSend >= bytesCount)
+                            {
+                                hasSend -= bytesCount;
+                                if (hasSend > 0)
+                                    continue;
 
-                            _outputBufferListSending.RemoveRange(0, i + 1);
+                                _outputBufferListSending.RemoveRange(0, i + 1);
+                                break;
+                            }
+                            // 已经发送的数据比数组中的少。
+                            ArraySegment<byte> segment = _outputBufferListSending[i];
+                            _outputBufferListSending[i] = segment.Slice(hasSend, segment.Count - hasSend);
+                            _outputBufferListSending.RemoveRange(0, i);
                             break;
                         }
-                        // 已经发送的数据比数组中的少。
-                        ArraySegment<byte> segment = _outputBufferListSending[i];
-                        _outputBufferListSending[i] = segment.Slice(hasSend, segment.Count - hasSend);
-                        _outputBufferListSending.RemoveRange(0, i);
-                        hasSend = 0;
-                        break;
                     }
-                    if (hasSend > 0)
-                        throw new Exception("hasSend too big.");
 
                     if (_outputBufferListSending.Count == 0)
                     {
                         _outputBufferListSending = _outputBufferList;
                         _outputBufferList = null;
+                        _outputBufferListSendingCountSum = _outputBufferListCountSum;
+                        _outputBufferListCountSum = 0;
                     }
                     else if (null != _outputBufferList)
                     {
                         _outputBufferListSending.AddRange(_outputBufferList);
                         _outputBufferList = null;
+                        _outputBufferListSendingCountSum = _outputBufferListCountSum;
+                        _outputBufferListCountSum = 0;
                     }
-                    if (null != _outputBufferListSending)
-                        _socket.BeginSend(_outputBufferListSending, System.Net.Sockets.SocketFlags.None, OnAsyncSend, this);
 
+                    if (null != _outputBufferListSending)
+                    {
+                        eventArgsSend.BufferList = _outputBufferListSending;
+                        if (false == Socket.SendAsync(eventArgsSend))
+                            ProcessSend(eventArgsSend);
+                    }
                     return;
                 }
                 
@@ -292,21 +356,13 @@ namespace Zeze.Net
                 {
                     _outputBufferListSending = _outputBufferList;
                     _outputBufferList = null;
-                    _socket.BeginSend(_outputBufferListSending, System.Net.Sockets.SocketFlags.None, OnAsyncSend, this);
-                }
-            }
-        }
+                    _outputBufferListSendingCountSum = _outputBufferListCountSum;
+                    _outputBufferListCountSum = 0;
 
-        private void OnAsyncSend(IAsyncResult ar)
-        {
-            try
-            {
-                int hasSend = _socket.EndSend(ar);
-                BeginSend(hasSend);
-            }
-            catch (Exception e)
-            {
-                Close(e);
+                    eventArgsSend.BufferList = _outputBufferListSending;
+                    if (false == Socket.SendAsync(eventArgsSend))
+                        ProcessSend(eventArgsSend);
+                }
             }
         }
 
@@ -322,8 +378,8 @@ namespace Zeze.Net
             {
                 try
                 {
-                    _socket?.Dispose();
-                    _socket = null;
+                    Socket?.Dispose();
+                    Socket = null;
                     Manager?.OnSocketClose(this, this.LastException);
                     Manager = null;
                 }
